% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apriori2df}
\alias{apriori2df}
\alias{tidy.rules}
\alias{tidy.itemsets}
\title{Tidy an \code{\link[arules:apriori]{arules::apriori()}} object}
\usage{
apriori2df(x)

\method{tidy}{rules}(x, ...)

\method{tidy}{itemsets}(x, ...)
}
\arguments{
\item{x}{An object of class \link[arules:rules-class]{arules::rules} or \link[arules:itemsets-class]{arules::itemsets} produced
by running \code{\link[arules:apriori]{arules::apriori()}}}

\item{...}{Unused, included for generic consistency only.}
}
\value{
a tibble (or data.frame) of the rules/itemsets
}
\description{
Converts output from \code{\link[arules:apriori]{arules::apriori()}} to a tibble/dataframe.
}
\details{
Use this instead of \code{\link[arules:inspect]{arules::inspect()}}, which only prints to screen.
}
\section{Methods (by generic)}{
\itemize{
\item \code{tidy}: Tidy a \link[arules:rules-class]{arules::rules} object

\item \code{tidy}: Tidy a \link[arules:itemsets-class]{arules::itemsets} object
}}

