% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{digits}
\alias{digits}
\title{Print exact decimal places}
\usage{
digits(x, k = 2)
}
\arguments{
\item{x}{Vector of numbers.}

\item{k}{Number of decimal places.}
}
\value{
A character vector of the number rounded to k decimal places.
}
\description{
Print exact decimal places
}
