% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homework.R
\name{homework}
\alias{homework}
\title{Rmarkdown template for homework}
\usage{
homework(nb = FALSE, ...)
}
\arguments{
\item{nb}{(logical) should an Rnotebook or html document be used}

\item{...}{other arguments passed to \code{html_format()}}
}
\description{
Rmarkdown template for homework
}
\seealso{
\url{https://rstudio.github.io/rstudio-extensions/rmarkdown_templates.html#custom-formats}
}
