% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6030.R
\docType{package}
\name{R6030}
\alias{R6030}
\title{R6018: A package for DS-6030}
\description{
The R6018 package is designed to help students in DS-6030. It provides some
helpful functions, formats, and datasets.
}
\section{R6030 functions}{

\code{\link{digits}}
}

\section{R6030 Formats}{

\code{\link{homework}}
}

\section{R6030 Datasets}{

TODO
}

